Move setting verification to command file
