+++ docker/entrypoint.sh
@@ -12,7 +12,7 @@ fi
 if [ "$IS_RQ" = 'yes' ]; then
   echo ">> Running database migrations."
   python manage.py migrate --no-input
-  echo ">> Verifying settings that require database migrations first."
+  echo ">> Verifying settings."
   python manage.py verify_settings
   echo ">> Starting RQ worker(s)"
